# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: License Key Delivery API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tenants/me:
    get:
      operationId: getMyTenant
      tags: [auth]
      summary: Get current tenant settings
      responses:
        "200":
          description: Tenant settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
    patch:
      operationId: updateMyTenant
      tags: [auth]
      summary: Update current tenant settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"

  # Products (admin)
...
  /orders/{id}/fulfill:
    post:
      operationId: fulfillOrder
      tags: [orders]
      summary: Manually fulfill a pending/failed order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Fulfilled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
...
  /stripe/checkout:
    post:
      operationId: createCheckoutSession
      tags: [stripe]
      summary: Create a Stripe Checkout session for a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "200":
          description: Checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResult"
        "404":
          description: Product not found
        "400":
          description: Product has no Stripe Price ID
...
components:
  schemas:
...
    Order:
      type: object
      required: [id, source, status, buyerEmail, productId, createdAt]
      properties:
        id:
          type: integer
        source:
          type: string
          enum: [manual, ebay, stripe]
        status:
          type: string
          enum: [pending, paid, fulfilled, failed, refunded]
        buyerEmail:
          type: string
        buyerName:
          type: ["string", "null"]
        productId:
          type: integer
        productName:
          type: ["string", "null"]
        quantity:
          type: integer
        ebayOrderId:
          type: ["string", "null"]
        ebayLineItemId:
          type: ["string", "null"]
        stripePaymentIntentId:
          type: ["string", "null"]
        stripeSessionId:
          type: ["string", "null"]
        assignedKeyId:
          type: ["integer", "null"]
        assignedKeyValue:
          type: ["string", "null"]
        failureReason:
          type: ["string", "null"]
        fulfilledAt:
          type: ["string", "null"]
        createdAt:
          type: string
...
    CheckoutInput:
      type: object
      required: [productId]
      properties:
        productId:
          type: integer
        successUrl:
          type: string
        cancelUrl:
          type: string
    CheckoutResult:
      type: object
      required: [url]
      properties:
        url:
          type: string
...
    Tenant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        status:
          type: string
        stripeSecretKey:
          type: string
          nullable: true
        stripeWebhookSecret:
          type: string
          nullable: true
        resendApiKey:
          type: string
          nullable: true
        fromEmail:
          type: string
          nullable: true
        supportEmail:
          type: string
          nullable: true
    TenantUpdate:
      type: object
      properties:
        name:
          type: string
        stripeSecretKey:
          type: string
        stripeWebhookSecret:
          type: string
        resendApiKey:
          type: string
        fromEmail:
          type: string
        supportEmail:
          type: string
...