# lib/api-zod/src/generated/types/tenant.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * License Key Delivery API
 * OpenAPI spec version: 0.2.0
 */

export interface Tenant {
  id?: number;
  name?: string;
  slug?: string;
  status?: string;
  stripeSecretKey?: string | null;
  stripeWebhookSecret?: string | null;
  resendApiKey?: string | null;
  fromEmail?: string | null;
  supportEmail?: string | null;
}